vision/src/main/python/sciencebirds/requirements.txt
cycler==0.10.0
decorator==4.4.2
FlattenList==0.1.4
imageio==2.8.0
imutils==0.5.3
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.5
opencv-contrib-python==4.4.0.40
opencv-python==4.4.0.40
pandas==1.1.1
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.6.3
